requirements.txt
beautifulsoup4==4.8.2
pandas==0.25.3
python-dateutil==2.8.1
regex==2020.2.20
requests==2.22.0
soupsieve==1.9.5
